+++ Slotmachine/src/main/java/com/example/casino/dto/stats/IUserStatsResponse.java
@@ -1,11 +1,13 @@
 package com.example.casino.dto.stats;
 
+import java.math.BigDecimal;
+
 public interface IUserStatsResponse {
     long getClientId();
     int getTotalGamesCount();
     int getTotalWinnings();
     int getTotalLosses();
-    double getTotalClientProfit();
-    double getTotalTurnoverFromClient();
-    double getTotalProfitFromClient();
+    BigDecimal getTotalClientProfit();
+    BigDecimal getTotalTurnoverFromClient();
+    BigDecimal getTotalProfitFromClient();
 }
