+++ Slotmachine/src/main/java/com/example/casino/dto/transaction/BankingTransactionRequest.java
@@ -3,17 +3,19 @@ package com.example.casino.dto.transaction;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Request to make transaction with banking service")
 public class BankingTransactionRequest implements IBankingTransactionRequest {
     @JsonProperty("invoicing_party")
     private String invoicing_party;
 
     @JsonProperty("amount")
-    private double amount;
+    private BigDecimal amount;
 
     public BankingTransactionRequest() {}
 
-    public BankingTransactionRequest(String invoicing_party, double amount) {
+    public BankingTransactionRequest(String invoicing_party, BigDecimal amount) {
         this.invoicing_party = invoicing_party;
         this.amount = amount;
     }
@@ -27,11 +29,11 @@ public class BankingTransactionRequest implements IBankingTransactionRequest {
         this.invoicing_party = invoicing_party;
     }
     @Override
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
     @Override
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
